+++ src/analyze/+mag/@IMAPTestingAnalysis/loadIALiRTData.m
@@ -39,6 +39,10 @@ function loadIALiRTData(this, primaryMetaData, secondaryMetaData)
         primary = renamevars(primary, ["x_pri", "y_pri", "z_pri", "rng_pri", "pri_coarse", "pri_fine"], newVariableNames);
         secondary = renamevars(secondary, ["x_sec", "y_sec", "z_sec", "rng_sec", "sec_coarse", "sec_fine"], newVariableNames);
 
+        % Add quality flag.
+        primary.quality = true(height(primary), 1);
+        secondary.quality = true(height(secondary), 1);
+
         % Current file meta data.
         pmd = primaryMetaData.copy();
         smd = secondaryMetaData.copy();