+++ src/analyze/+mag/@IMAPTestingAnalysis/loadScienceData.m
@@ -33,6 +33,10 @@ function loadScienceData(this, primaryMetaData, secondaryMetaData)
         primary = renamevars(primary, ["x_pri", "y_pri", "z_pri", "rng_pri", "pri_coarse", "pri_fine"], newVariableNames);
         secondary = renamevars(secondary, ["x_sec", "y_sec", "z_sec", "rng_sec", "sec_coarse", "sec_fine"], newVariableNames);
 
+        % Add quality flag.
+        primary.quality = true(height(primary), 1);
+        secondary.quality = true(height(secondary), 1);
+
         % Current file meta data.
         [mode, primaryFrequency, secondaryFrequency, packetFrequency] = extractFileMetaData(this.ScienceFileNames(i));